Me atasco con un error sin nada en rojo

+++ src/descargas/Utils.java
@@ -2,7 +2,7 @@ package descargas;
 
 public class Utils {
 
-	public static float getTamano(Fichero[] listaDescargas) {
+	public static double getTamano(Fichero[] listaDescargas) {
 		float resultado  = 0;
 		for (int i = 0; i < listaDescargas.length; i++) {
 			resultado += listaDescargas[i].getTamano();
@@ -10,7 +10,7 @@ public class Utils {
 		return resultado;
 	}
 
-	public static float getTiempo(Fichero[] listaDescargas, Conexion conexion) {
+	public static double getTiempo(Fichero[] listaDescargas, Conexion conexion) {
 		return getTamano(listaDescargas)/conexion.getVelocidadDescarga();
 	}
